refactor: move basic ops to dedicated module
